Add validation

+++ CommonBeneficiary.Application/Features/RelationTypes/Handlers/Commands/DeleteRelationTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using CommonBeneficiary.Application.Contracts.Persistance;
+using CommonBeneficiary.Application.Core.Responses;
 using CommonBeneficiary.Application.Features.RelationTypes.Requests.Commands;
 using MediatR;
 using System;
@@ -10,7 +11,7 @@ using System.Threading.Tasks;
 
 namespace CommonBeneficiary.Application.Features.RelationTypes.Handlers.Commands
 {
-    public class DeleteRelationTypeCommandHandler : IRequestHandler<DeleteRelationTypeCommand>
+    public class DeleteRelationTypeCommandHandler : IRequestHandler<DeleteRelationTypeCommand, BaseResponse<Unit>>
     {
         private readonly IRelationTypeRepository _repository;
         private readonly IMapper _mapper;
@@ -21,11 +22,19 @@ namespace CommonBeneficiary.Application.Features.RelationTypes.Handlers.Commands
             _mapper = mapper;
         }
 
-        public async Task<Unit> Handle(DeleteRelationTypeCommand request, CancellationToken cancellationToken)
+        public async Task<BaseResponse<Unit>> Handle(DeleteRelationTypeCommand request, CancellationToken cancellationToken)
         {
             var relationType = await _repository.Get(request.Id);
-            await _repository.Delete(relationType);
-            return Unit.Value;
+            if(relationType == null)
+                return BaseResponse<Unit>.Failure(errors: new List<string> { "رابطه با کد داده شده یافت نشد" });
+
+            var result = await _repository.Delete(relationType);
+            if (result > 0)
+                return BaseResponse<Unit>.Success();
+            else
+                return BaseResponse<Unit>.Failure(errors: new List<string> { "خطا در حذف رابطه" });
+
+
         }
     }
 }
